find using location and fix feedback

+++ src/main/java/enclave/encare/encare/model/Hospital.java
@@ -18,8 +18,8 @@ public class Hospital {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long hospitalId;
     private String description;
-    private long latMap;
-    private long longMap;
+    private double latMap;
+    private double longMap;
     private long rating;
     private long countRating;
     private String address;
